SurgeScript now uses 64-bit floating point numbers

+++ src/surgescript/compiler/codegen.h
@@ -105,7 +105,7 @@ void emit_state(surgescript_nodecontext_t context);
 void emit_identifier(surgescript_nodecontext_t context, const char* identifier, int line);
 void emit_null(surgescript_nodecontext_t context);
 void emit_bool(surgescript_nodecontext_t context, bool value);
-void emit_number(surgescript_nodecontext_t context, float value);
+void emit_number(surgescript_nodecontext_t context, double value);
 void emit_string(surgescript_nodecontext_t context, const char* value);
 void emit_object(surgescript_nodecontext_t context, unsigned handle);
 void emit_zero(surgescript_nodecontext_t context);
