+++ src/surgescript/runtime/sslib/gc.c
@@ -25,7 +25,7 @@
 #include "../../util/util.h"
 
 /* constants */
-static const float DEFAULT_GC_INTERVAL = 1.0f; /* will run GC.collect() every DEFAULT_GC_INTERVAL seconds (by default) */
+static const double DEFAULT_GC_INTERVAL = 1.0; /* will run GC.collect() every DEFAULT_GC_INTERVAL seconds (by default) */
 
 /* private stuff */
 static surgescript_var_t* fun_constructor(surgescript_object_t* object, const surgescript_var_t** param, int num_params);
@@ -63,7 +63,7 @@ void surgescript_sslib_register_gc(surgescript_vm_t* vm)
 surgescript_var_t* fun_constructor(surgescript_object_t* object, const surgescript_var_t** param, int num_params)
 {
     surgescript_heap_t* heap = surgescript_object_heap(object);
-    float now = surgescript_util_gettickcount() * 0.001f;
+    double now = surgescript_util_gettickcount() * 0.001;
 
     ssassert(INTERVAL_ADDR == surgescript_heap_malloc(heap));
     ssassert(LASTCOLLECT_ADDR == surgescript_heap_malloc(heap));
@@ -79,9 +79,9 @@ surgescript_var_t* fun_main(surgescript_object_t* object, const surgescript_var_
 {
     surgescript_objectmanager_t* manager = surgescript_object_manager(object);
     surgescript_heap_t* heap = surgescript_object_heap(object);
-    float interval = surgescript_var_get_number(surgescript_heap_at(heap, INTERVAL_ADDR));
-    float last_collect = surgescript_var_get_number(surgescript_heap_at(heap, LASTCOLLECT_ADDR));
-    float now = surgescript_util_gettickcount() * 0.001f;
+    double interval = surgescript_var_get_number(surgescript_heap_at(heap, INTERVAL_ADDR));
+    double last_collect = surgescript_var_get_number(surgescript_heap_at(heap, LASTCOLLECT_ADDR));
+    double now = surgescript_util_gettickcount() * 0.001;
 
     surgescript_objectmanager_garbagecheck(manager);
     if(now - last_collect >= interval) {
@@ -125,7 +125,7 @@ surgescript_var_t* fun_getinterval(surgescript_object_t* object, const surgescri
 surgescript_var_t* fun_setinterval(surgescript_object_t* object, const surgescript_var_t** param, int num_params)
 {
     surgescript_heap_t* heap = surgescript_object_heap(object);
-    float new_interval = ssmax(0.0f, surgescript_var_get_number(param[0]));
+    double new_interval = ssmax(0.0, surgescript_var_get_number(param[0]));
     surgescript_var_set_number(surgescript_heap_at(heap, INTERVAL_ADDR), new_interval);
     return NULL;
 }
